# api/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Self-hosted media server for movies, TV shows, and music",
        "title": "ViewRA Media Server API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "ViewRA Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "0.0.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/movies": {
            "get": {
                "description": "Returns a list of all movies in a specific library with optional pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "List movies",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Library ID to filter movies",
                        "name": "library_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Number of items per page (default: 50, max: 200)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items to skip (default: 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_viewra_viewra_internal_application_movies.ListMoviesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/movies/search": {
            "get": {
                "description": "Searches for movies by title in a specific library with optional pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Search movies",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Library ID to search in",
                        "name": "library_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Search query (title)",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Number of items per page (default: 50, max: 200)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items to skip (default: 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_viewra_viewra_internal_application_movies.ListMoviesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/movies/{id}": {
            "get": {
                "description": "Returns details of a specific movie including all metadata",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Get a movie by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Movie ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_viewra_viewra_internal_application_movies.MovieResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_viewra_viewra_internal_application_movies.ListMoviesResponse": {
            "type": "object",
            "properties": {
                "movies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_viewra_viewra_internal_application_movies.MovieResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/github_com_viewra_viewra_internal_domain_common.PaginationMetadata"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "github_com_viewra_viewra_internal_application_movies.MovieResponse": {
            "type": "object",
            "required": [
                "file_path",
                "id",
                "library_id",
                "title"
            ],
            "properties": {
                "audio_codec": {
                    "type": "string"
                },
                "awards_summary": {
                    "type": "string"
                },
                "bitrate": {
                    "type": "integer"
                },
                "budget": {
                    "type": "integer"
                },
                "cast": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "container_format": {
                    "type": "string"
                },
                "content_advisories": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "content_rating": {
                    "type": "string"
                },
                "country_of_origin": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "director": {
                    "type": "string"
                },
                "duration": {
                    "description": "in seconds",
                    "type": "integer"
                },
                "file_path": {
                    "type": "string"
                },
                "file_size": {
                    "type": "integer"
                },
                "frame_rate": {
                    "type": "number"
                },
                "genre": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "height": {
                    "type": "integer"
                },
                "id": {
                    "description": "Base Media fields",
                    "type": "integer"
                },
                "imdb_id": {
                    "type": "string"
                },
                "is_extra": {
                    "type": "boolean"
                },
                "library_id": {
                    "type": "integer"
                },
                "maturity_rating": {
                    "type": "integer"
                },
                "original_language": {
                    "type": "string"
                },
                "original_title": {
                    "type": "string"
                },
                "plot": {
                    "type": "string"
                },
                "release_date": {
                    "type": "string"
                },
                "revenue": {
                    "type": "integer"
                },
                "runtime_minutes": {
                    "type": "integer"
                },
                "sort_title": {
                    "type": "string"
                },
                "tagline": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "tmdb_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "video_codec": {
                    "type": "string"
                },
                "width": {
                    "description": "Technical metadata",
                    "type": "integer"
                },
                "year": {
                    "description": "Movie-specific fields",
                    "type": "integer"
                }
            }
        },
...
        "github_com_viewra_viewra_internal_domain_common.PaginationMetadata": {
            "type": "object",
            "properties": {
                "hasMore": {
                    "description": "Whether there are more items available",
                    "type": "boolean"
                },
                "limit": {
                    "description": "Number of items per page",
                    "type": "integer"
                },
                "offset": {
                    "description": "Number of items skipped",
                    "type": "integer"
                },
                "total": {
                    "description": "Total number of items across all pages",
                    "type": "integer"
                }
            }
        },
...
        "internal_api_handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}